Move LatestAttestations to block_import mod
